test(travis): Update MODPATH 7 autotest
Update travis test to run flow model prior to running MODPATH 7. Add
MODFLOW-USG simulations, change case of MODFLOW-2005 namefile
filenames to lower case for linux OS, fix a few errors in the
distributed example files, and build mf2005, mfusg, and mf6 as part of
MODPATH 7 build.

+++ autotest/t007_test.py
@@ -187,12 +187,16 @@ def run_modpath7(fn):
 
 
 def clean_up(pth, exe):
-    # clean up
-    print('Removing folder ' + pth)
-    shutil.rmtree(pth)
+    # clean up downloaded directories
+    pth = os.path.join(dstpth, pth)
+    if os.path.isdir(pth):
+        print('Removing folder ' + pth)
+        shutil.rmtree(pth)
+    # clean up compiled executables
     binpth = os.path.join(dstpth, exe)
-    print('Removing ' + binpth)
-    os.remove(binpth)
+    if os.path.isfile(binpth):
+        print('Removing ' + binpth)
+        os.remove(binpth)
     return
